LF-4625 Move functionality to own route and controller
